+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Attribute 'locked' from MetadataSchemaRecord changed to 'doNotSync' (issue #37)
 - Change in related schema and data (add identifier type to identifier)
 - Store all identifiers as global identifiers (type INTERNAL -> type URL)
+- For registering a schema mimetype is no longer mandatory.
 ### Fixed
 - Filtering metadata documents by resourceId, schemaId
 - Filtering schema documents by mimetype